Proper count of Objective Function

+++ src/SimplexTable.hpp
@@ -17,7 +17,7 @@ private:
     int numberOfColumns = {0};
     int numberOfLines = {0};
 
-    int valueOfObjectiveFunction = {0};
+    float valueOfObjectiveFunction = {0};
     int numberOfIteration = {0};
 
     std::vector<std::vector<float>> simplexTable;
